Revert "Preliminary threat processing integration"
This reverts commit 14f1cf437055a96eec977c760ff610476822fbaa.
