+++ .secrets.baseline
@@ -213,15 +213,6 @@
         "is_secret": false
       }
     ],
-    "packages/document-service/.jest/test.env.js": [
-      {
-        "type": "Secret Keyword",
-        "filename": "packages/document-service/.jest/test.env.js",
-        "hashed_secret": "fee7c63caf4bbe9083d881287ee4d75e7c00ce2f",
-        "is_verified": false,
-        "line_number": 2
-      }
-    ],
     "packages/geoprocessing-service/.jest/test.env.js": [
       {
         "type": "Base64 High Entropy String",
@@ -239,16 +230,7 @@
         "line_number": 3,
         "is_secret": false
       }
-    ],
-    "packages/utils-lib/src/helpers/__tests__/get-bearer-token-spec.js": [
-      {
-        "type": "Secret Keyword",
-        "filename": "packages/utils-lib/src/helpers/__tests__/get-bearer-token-spec.js",
-        "hashed_secret": "fee7c63caf4bbe9083d881287ee4d75e7c00ce2f",
-        "is_verified": false,
-        "line_number": 8
-      }
     ]
   },
-  "generated_at": "2022-08-01T10:28:01Z"
+  "generated_at": "2022-07-21T10:19:53Z"
 }
