+++ packages/application-to-register-functions/package.json
@@ -20,7 +20,6 @@
   "dependencies": {
     "@azure/storage-blob": "^12.9.0",
     "@defra/bng-connectors-lib": "0.0.1",
-    "@defra/bng-document-service": "^0.0.1",
     "@defra/bng-errors-lib": "^0.0.1",
     "@defra/bng-geoprocessing-service": "0.0.1",
     "defra-logging-facade": "git+https://github.com/DEFRA/defra-logging-facade.git#bb51e5b"
