--- packages/document-service/.jest/test.env.js
@@ -1,8 +1,0 @@
-process.env.AV_API_CLIENT_ID = 'mockClientId',
-process.env.AV_API_CLIENT_SECRET = 'mockClientSecret',
-process.env.AV_API_SCOPE = 'mockScope',
-process.env.AV_API_OCP_APIM_SUBSCRIPTION_KEY = 'mockSubscriptionKey'
-process.env.AV_API_USER_ID = 'mockUserId',
-process.env.AV_API_BASE_URL = 'mockBaseUrl',
-process.env.AV_API_TOKEN_URL = 'mockTokenUrl'
-process.env.AV_API_RESULT_RETRIEVAL_INTERVAL_MS = 500
