--- packages/document-service/jest.config.cjs
@@ -1,9 +1,0 @@
-module.exports = {
-  rootDir: '.',
-  moduleNameMapper: {
-    '@defra/bng-connectors-lib/azure-storage': '<rootDir>/node_modules/@defra/bng-azure-storage-test-utils/node_modules/@defra/bng-connectors-lib/src/helpers/azure-storage.js'
-  },
-  setupFiles: ['<rootDir>/.jest/test.env.js'],
-  setupFilesAfterEnv: ['<rootDir>/.jest/setup.js'],
-  testPathIgnorePatterns: ['__tests__/helpers']
-}
