+++ packages/document-service/package.json
@@ -14,13 +14,5 @@
     "directory": "document-service"
   },
   "author": "Department for Environment, Food and Rural Affairs <https://www.gov.uk/government/organisations/department-for-environment-food-rural-affairs>",
-  "license": "OGL-UK-3.0",
-  "dependencies": {
-    "@defra/bng-utils-lib": "^0.0.1",
-    "axios": "^0.27.2"
-  },
-  "devDependencies": {
-    "@defra/bng-azure-storage-test-utils": "0.0.1"
-  }
-
+  "license": "OGL-UK-3.0"
 }
