--- packages/document-service/src/__mock-data__/uploads/mock-data.json
@@ -1,3 +1,0 @@
-{
-  "key": "value"
-}
