+++ packages/document-service/src/service.js
@@ -1,4 +1,18 @@
-import screenDocumentForThreats from './helpers/screen-document.js'
-import upload from './helpers/upload-document.js'
+// A document upload function with configuration driven implementation.
+const upload = async (logger, config, document) => {
+  const uploadConfig = createUploadConfiguration(logger, config)
+  // Delegate upload processing to the function referenced in the configuration.
+  // The configured funcion encapsulates all implementation specific details.
+  return config.functionConfig.uploadFunction(logger, uploadConfig, document)
+}
 
-export { screenDocumentForThreats, upload }
+const createUploadConfiguration = (logger, config) => {
+  // Clone the original configuration.
+  const uploadConfig = JSON.parse(JSON.stringify(config))
+  // Delete all configuration that does not need to be passed to the configured functions.
+  delete uploadConfig.functionConfig
+  logger.log('Upload configuration has been created')
+  return uploadConfig
+}
+
+export { upload }
