--- packages/errors-lib/__tests__/threat-screening-error-spec.js
@@ -1,13 +1,0 @@
-import { ThreatScreeningError } from '../src/errors.js'
-
-const threatScreeningDetails = {}
-
-describe('Threat screening errors', () => {
-  it('should supplement a standard Error with threat screening failure details', async () => {
-    try {
-      throw new ThreatScreeningError(threatScreeningDetails)
-    } catch (err) {
-      expect(err.threatScreeningDetails).toStrictEqual(threatScreeningDetails)
-    }
-  })
-})
