+++ packages/errors-lib/src/errors.js
@@ -1,6 +1,5 @@
 import CoordinateSystemValidationError from './coordinate-system-validation-error.js'
-import ThreatScreeningError from './threat-screening-error.js'
 import UploadTypeValidationError from './upload-type-validation-error.js'
 import ValidationError from './validation-error.js'
 
-export { CoordinateSystemValidationError, ThreatScreeningError, UploadTypeValidationError, ValidationError }
+export { CoordinateSystemValidationError, UploadTypeValidationError, ValidationError }
