--- packages/errors-lib/src/threat-screening-error.js
@@ -1,8 +1,0 @@
-// Adapted from https://medium.com/@xjamundx/custom-javascript-errors-in-es6-aa891b173f87
-export default class ThreatScreeningError extends Error {
-  constructor (threatScreeningDetails, ...args) {
-    super(...args)
-    this.threatScreeningDetails = threatScreeningDetails
-    Error.captureStackTrace(this, ThreatScreeningError)
-  }
-}
