--- packages/utils-lib/README.md
@@ -1,3 +1,0 @@
-# Utilities Library
-
-A library of utilities for the Biodiversity Net Gain Service.
