--- packages/utils-lib/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "@defra/bng-utils-lib",
-  "version": "0.0.1",
-  "type": "module",
-  "private": false,
-  "description": "A library of utilities for the Biodiversity Net Gain Service",
-  "main": "src/utils.js",
-  "scripts": {
-    "test": "echo \\\"Error: run tests from root\\\" && exit 1"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/DEFRA/biodiversity-net-gain-service.git"
-  },
-  "author": "Department for Environment, Food and Rural Affairs <https://www.gov.uk/government/organisations/department-for-environment-food-rural-affairs>",
-  "license": "OGL-UK-3.0",
-  "dependencies": {
-    "axios": "^0.27.2"
-  }
-}
