--- packages/utils-lib/src/utils.js
@@ -1,3 +1,0 @@
-import getBearerToken from './helpers/get-bearer-token.js'
-
-export { getBearerToken }
